pyproject.toml
[tool.poetry]
name = "bridget"
version = "0.1.0"
description = ""
authors = ["Ginger <75683114+gingerindustries@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sechat = {git = "https://github.com/GingerIndustries/sechat", rev = "v2"}
"discord.py" = "^2.3.2"
bs4 = "^0.0.1"
markdownify = "^0.11.6"
odmantic = "^0.9.2"
pydantic = "<2.0"
discord-markdown-ast-parser = "^1.0.6"
lxml = "^4.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
